--- lib/uploadImagesToCloudinary.js
@@ -1,12 +1,0 @@
-const cloudinary = require('cloudinary');
-
-
-cloudinary.config({ 
-    cloud_name: 'sample', 
-    api_key: '874837483274837', 
-    api_secret: 'a676b67565c6767a6767d6767f676fe1',
-    secure: true
-  });
-
-
- cons
